Create TransactionListItemRow
